Move test util SpyMediaSession to own class
