+++ code/edgemedia/src/test/java/com/adobe/marketing/mobile/edge/media/internal/MediaSessionTests.kt
@@ -127,63 +127,4 @@ class MediaSessionTests {
         assertFalse(session.handleSessionAbortCalled)
         assertNull(session.handleSessionAbortParamAbortHandler)
     }
-
-    private class SpyMediaSession(
-        id: String,
-        state: MediaState,
-        dispatchHandler: (Event) -> Unit
-    ) : MediaSession(
-        id,
-        state,
-        dispatchHandler
-    ) {
-
-        var handleMediaStateUpdateCalled: Boolean = false
-        override fun handleMediaStateUpdate() {
-            handleMediaStateUpdateCalled = true
-        }
-
-        var handleSessionEndCalled: Boolean = false
-        var handleSessionEndParamEndHandler: (() -> Unit)? = null
-        override fun handleSessionEnd(sessionEndHandler: () -> Unit) {
-            handleSessionEndCalled = true
-            handleSessionEndParamEndHandler = sessionEndHandler
-        }
-
-        var handleSessionAbortCalled: Boolean = false
-        var handleSessionAbortParamAbortHandler: (() -> Unit)? = null
-        override fun handleSessionAbort(sessionAbortHandler: () -> Unit) {
-            handleSessionAbortCalled = true
-            handleSessionAbortParamAbortHandler = sessionAbortHandler
-        }
-
-        var handleQueueEventCalled: Boolean = false
-        var handleQueueEventParamXDMMediaEvent: XDMMediaEvent? = null
-        override fun handleQueueEvent(event: XDMMediaEvent) {
-            handleQueueEventCalled = true
-            handleQueueEventParamXDMMediaEvent = event
-        }
-
-        var handleSessionUpdateCalled: Boolean = false
-        var handleSessionUpdateParamRequestEventId: String? = null
-        var handleSessionUpdateParamBackendSessionId: String? = null
-        var handleSessionUpdateParamAbortHandler: (() -> Unit)? = null
-        override fun handleSessionUpdate(requestEventId: String, backendSessionId: String?, sessionAbortHandler: () -> Unit) {
-            handleSessionUpdateCalled = true
-            handleSessionUpdateParamRequestEventId = requestEventId
-            handleSessionUpdateParamBackendSessionId = backendSessionId
-            handleSessionUpdateParamAbortHandler = sessionAbortHandler
-        }
-
-        var handleErrorResponseCalled: Boolean = false
-        var handleErrorResponseParamRequestEventId: String? = null
-        var handleErrorResponseParamData: Map<String, Any>? = null
-        var handleErrorResponseParamAbortHandler: (() -> Unit)? = null
-        override fun handleErrorResponse(requestEventId: String, data: Map<String, Any>, sessionAbortHandler: () -> Unit) {
-            handleErrorResponseCalled = true
-            handleErrorResponseParamRequestEventId = requestEventId
-            handleErrorResponseParamData = data
-            handleErrorResponseParamAbortHandler = sessionAbortHandler
-        }
-    }
 }
